feat(alerts): include artist series in suggested filters (#5471)
* refactor: extract previewSavedSearch files to dedicated dir

* refactor: extract generateDisplayName

* refactor: extract resolveHref

* refactor: extract suggestedFilters

* feat: add artist series suggestions based on artist aggs

* feat: allow field for alert source (artwork vs artist)

* feat: add artist series suggestions based on artwork attributes
