+++ _schemaV2.graphql
@@ -323,6 +323,20 @@ type AlertsCounts {
   totalUserSearchCriteriaCount: Int
 }
 
+input AlertSource {
+  # The database id of the object from which the alert originates
+  id: ID
+
+  # The type of object from which the alert originates
+  type: AlertSourceType
+}
+
+# The context from which the alert originates
+enum AlertSourceType {
+  ARTIST
+  ARTWORK
+}
+
 # A connection to a list of items.
 type AlertsSummaryArtistConnection {
   # A list of edges.
@@ -14507,7 +14521,10 @@ type PreviewSavedSearch {
   labels: [SearchCriteriaLabel]!
 
   # Suggested filters for the user to use based on their search criteria
-  suggestedFilters: [SearchCriteriaLabel!]!
+  suggestedFilters(
+    # The context from which the alert originates
+    source: AlertSource
+  ): [SearchCriteriaLabel!]!
 }
 
 input PreviewSavedSearchAttributes {
@@ -15914,6 +15931,7 @@ type Query {
 
   # A previewed saved search
   previewSavedSearch(
+    # The criteria which describe the alert
     attributes: PreviewSavedSearchAttributes
   ): PreviewSavedSearch
 
@@ -20540,6 +20558,7 @@ type Viewer {
 
   # A previewed saved search
   previewSavedSearch(
+    # The criteria which describe the alert
     attributes: PreviewSavedSearchAttributes
   ): PreviewSavedSearch
