+++ src/schema/v2/schema.ts
@@ -202,7 +202,7 @@ import { PhoneNumber } from "./phoneNumber"
 import { PreviewSavedSearchField } from "./previewSavedSearch"
 import { Profiles } from "./profiles"
 import { RecentlySoldArtworks } from "./recentlySoldArtworks"
-import { SearchCriteriaLabel } from "./searchCriteriaLabel"
+import { SearchCriteriaLabel } from "./previewSavedSearch/searchCriteriaLabel"
 import { shortcut } from "./shortcut"
 import { Users } from "./users"
 import { createUserInterestForUser } from "./users/createUserInterestForUserMutation"
